refactor(origin-alloc): move #[global_allocator] out of lib; tighten thread-arena types + check epoch mallctl rc
Bundles three fixes that all live in `crates/origin-alloc/`:

Important 3 — Remove the lib-level `#[global_allocator]` declaration
that forced every binary linking `origin-alloc` (with the `jemalloc`
feature) to use jemalloc as the global allocator. Library crates
should not own that decision; it's a bin's job. Now:

  - The lib re-exports `tikv_jemallocator::Jemalloc as
    JemallocAllocator` (cfg `jemalloc`) so bins can opt in with their
    own `#[global_allocator]` line.
  - The integration test in `tests/reset_destroy.rs` declares its own
    `#[global_allocator]` so `allocated_bytes` still moves in response
    to `Vec` allocations during the test.
  - Per-arena MALLCTL calls keep working regardless of the global
    allocator choice, since `tikv-jemalloc-sys` still links jemalloc
    symbols in.

Important 5 — `snapshot()` ignored the return code from the `epoch`
mallctl, so a stats-refresh failure would silently return stale
counters. Now bind `ret` and `assert_eq!(ret, 0, ...)`, with an
expanded SAFETY comment.

Important 6 — Tightened the thread-arena type from `Option<usize>` to
`Option<u32>` end-to-end so the `u32::try_from(...).expect(...)` site
in `restore_thread_arena` disappears. This is the larger of the two
fixes the reviewer offered: it touches `jemalloc_backend.rs` (return
type), `noop_backend.rs` (mirror to `Option<u32>`), and `scope.rs`
(`ArenaScope::prev_index` typed `Option<u32>`). `with_arena` and
`Drop for ArenaScope` flow through unchanged because they were
already type-polymorphic at the call site.

Tests: cargo test -p origin-alloc passes (jemalloc-gated tests no-op
on this Windows host as expected). Clippy clean with `-D warnings`.

+++ crates/origin-alloc/src/scope.rs
@@ -9,8 +9,9 @@ use crate::arena_id::ArenaId;
 pub struct ArenaScope {
     id: ArenaId,
     // Restoration of the prior thread-arena binding is the backend's job; this
-    // field is private and lives only for `Drop`.
-    pub(crate) prev_index: Option<usize>,
+    // field is private and lives only for `Drop`. Stored as `Option<u32>` to
+    // match the jemalloc backend's native arena index type.
+    pub(crate) prev_index: Option<u32>,
 }
 
 impl ArenaScope {
@@ -20,7 +21,7 @@ impl ArenaScope {
         self.id
     }
 
-    pub(crate) const fn new(id: ArenaId, prev_index: Option<usize>) -> Self {
+    pub(crate) const fn new(id: ArenaId, prev_index: Option<u32>) -> Self {
         Self { id, prev_index }
     }
 }
